Se renombraron modelos relacionados a predios, se agrego relación polimorfica de terrenoscomun a predios, se finalizo área de notificación, se finalizó área de impresión, se comenzo, área de ficha técnica

+++ database/migrations/2023_03_28_151911_create_tramites_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('estado');
             $table->string("tipo_tramite");
-            $table->unsignedBigInteger('folio')->unique();
+            $table->unsignedInteger('folio')->unique();
             $table->foreignId('servicio_id')->constrained();
             $table->string('solicitante');
             $table->unsignedDecimal("monto", 18, 2);
